Fix 2nd equals click not showing the operator and 2nd operand
- as usual when the equals buttons is pressed on applying an operator to two
operands, and the equals button is pressed a second time on the result
(just the result [single operand, no operator]), the previous operator
and 2nd operand is carried on in the next operator to apply themm again.
This was still happening as expected but the history was not showing the
second operand and the operator. Fixed it.